+++ R/defunct-colourInput.R
@@ -1,0 +1,20 @@
+#' @rdname shinyjs-defunct
+#' @export
+colourInput <- function(...) {
+  .Defunct("colourInput", package = "colourpicker",
+           msg = "colourInput() has been moved to the 'colourpicker' package.")
+}
+
+#' @rdname shinyjs-defunct
+#' @export
+updateColourInput <- function(...) {
+  .Defunct("updateColourInput", package = "colourpicker",
+           msg = "updateColourInput() has been moved to the 'colourpicker' package.")
+}
+
+#' @rdname shinyjs-defunct
+#' @export
+colourPicker <- function(...) {
+  .Defunct("colourPicker", package = "colourpicker",
+           msg = "colourPicker() has been moved to the 'colourpicker' package.")
+}